Tidying things

+++ elia_chat/__main__.py
@@ -4,10 +4,12 @@ Elia CLI
 
 import asyncio
 import pathlib
+from textwrap import dedent
 import tomllib
 from typing import Tuple
 
 import click
+from rich.console import Console
 
 from elia_chat.app import Elia
 from elia_chat.config import LaunchConfig
@@ -15,6 +17,8 @@ from elia_chat.database.import_chatgpt import import_chatgpt_data
 from elia_chat.database.database import create_database, sqlite_file_name
 from elia_chat.launch_args import QuickLaunchArgs
 from elia_chat.locations import config_file
+
+console = Console()
 
 
 @click.group(invoke_without_command=True)
@@ -50,9 +54,28 @@ def reset() -> None:
     This command will delete the database file and recreate it.
     Previously saved conversations and data will be lost.
     """
-    sqlite_file_name.unlink(missing_ok=True)
-    asyncio.run(create_database())
-    click.echo(f"♻️  Database reset @ {sqlite_file_name}")
+    from rich.padding import Padding
+    from rich.text import Text
+
+    console.print(
+        Padding(
+            Text.from_markup(
+                dedent(f"""\
+[u b red]Warning![/]
+
+[b red]This will delete all messages and chats.[/]
+
+You may wish to create a backup of \
+"[bold blue u]{str(sqlite_file_name.resolve().absolute())}[/]" before continuing.
+            """)
+            ),
+            pad=(1, 2),
+        )
+    )
+    if click.confirm("Delete all chats?", abort=True):
+        sqlite_file_name.unlink(missing_ok=True)
+        asyncio.run(create_database())
+        console.print(f"♻️  Database reset @ {sqlite_file_name}")
 
 
 @cli.command("import")
